pyproject.toml
[tool.poetry]
name = "variables-economicas"
version = "0.1.0"
description = "Conexión con API para variables económicas"
authors = ["nico <niconte3500@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pandas = "^2.2.3"
requests = "^2.32.3"
certifi = "^2024.8.30"
python-dotenv = "^1.0.1"
SQLAlchemy = ">=1.4.36,<2.0"
apache-airflow = "^2.10.2"
psycopg2 = "^2.9.10"
sqlalchemy-redshift = "^0.8.14"
pyarrow = "^17.0.0"
fastparquet = "^2024.5.0"
awswrangler = "^3.9.1"
redshift-connector = "^2.1.3"
plotly = "^5.24.1"
dash = "^2.18.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
